refactor: move useCounter to external composable file
